She would merengue
and do the chacha

So in this one I basically looked what was in that zip file, and put it into master. The main issue is that we didn't have a 3rd room, so I added a blank 3rd room that the cloud can go to.

I also added in movement that feels more like a cloud, so the arrow keys apply more of a force to the cloud. when you stop pressing a key the cloud will gluide a halt.